fix(app-webdir-ui): remove alpha filter option from faculty rank
